Add Correspondence API and Correct PaymentDetails
Add the Correspondence API
Add CorrespondenceDTOs
Add CorrespondenceRepository and interface
Add CorrespondenceService and interface

Add OrderId column to PaymentDetails table

Add GetUserPayments, paymentService.GetAllPaymentDetailsByUser methods in order to populate user purchase history in User Dashboard on the front end

The Correspondence API allows for a user to send inquiries to the admin of the site and to view their own Inquiry History
The admin will have an Inquiry queue which they can use to respond to Inquries.
When an admin responds to an inquiry there is an email notification fired off to the user. The admin is encouraged to provide an admin email address or a contact email if the inquiry will go beyond the initial communication.

Also corrected the paymentDetails table
Added OrderId which gives us the ID that is generated on the backend and displayed on the payment success page when the checkout has completed.